requirements.txt
djangorestframework
markdown
django-filter 
jwt
mysqlclient
